fixed sql errors and added seed data file

+++ Catacombs/Models/Movies.cs
@@ -9,7 +9,7 @@
         public bool watched { get; set; }
         public string poster_path { get; set; }
         public string overview { get; set; }
-        public int popularity { get; set; }
+        public decimal popularity { get; set; }
         public decimal vote_average { get; set; }
         public Users Users { get; set; }
     }
